requirements.txt
Django==1.11.16
django-extensions==2.1.3
pygraphviz==1.5
django-userena==2.0.1
django-mail-templated
